refactor(visio): TeacherVisioList sous 300 lignes — éclatement complet (H13)
871 -> 274 l. Logique -> useTeacherVisioList.js ; sections -> VisioStatsCards/VisioSeanceCard/VisioSeanceSection. Appels lmsService + cache 'teacher_visio' + buildJitsiUrl inchanges. CSS verbatim (@keyframes pulse duplique scoped). Tests parite verts + nouveaux. Dette preexistante conservee (formatDate/Time locaux, new Date non-ISO, error-state global).
